Some casts now not needed

+++ base/oix/imisc.r
@@ -474,7 +474,7 @@ int check_access(struct class_field *cf, struct b_class *instance_class)
  */
 int lookup_class_field_by_name(struct b_class *class, dptr name)
 {
-    int i, c, m, l = 0, r = (int)class->n_instance_fields + (int)class->n_class_fields - 1;
+    int i, c, m, l = 0, r = class->n_instance_fields + class->n_class_fields - 1;
     while (l <= r) {
         m = (l + r) / 2;
         i = class->sorted_fields[m];
@@ -495,7 +495,7 @@ int lookup_class_field_by_name(struct b_class *class, dptr name)
  */
 int lookup_class_field_by_fnum(struct b_class *class, word fnum)
 {
-    int i, m, l = 0, r = (int)class->n_instance_fields + (int)class->n_class_fields - 1;
+    int i, m, l = 0, r = class->n_instance_fields + class->n_class_fields - 1;
     word c;
     while (l <= r) {
         m = (l + r) / 2;
@@ -533,7 +533,7 @@ int lookup_class_field(struct b_class *class, dptr query, struct inline_field_ca
          * Check if we have a inline cache match.
          */
         if (ic->class == (union block *)class)
-            return (int)ic->index;
+            return ic->index;
 
         /*
          * Query is a field number (from an Op_field).
@@ -588,7 +588,7 @@ int lookup_class_field(struct b_class *class, dptr query, struct inline_field_ca
              */
             word i = cvpos(IntVal(*query), nf);
             if (i != CvtFail && i <= nf)
-                return (int)i - 1;
+                return i - 1;
             else
                 return -1;
         }
@@ -622,7 +622,7 @@ static int record_access(dptr cargp, struct inline_field_cache *ic)
 
 int lookup_record_field_by_name(struct b_constructor *recdef, dptr name)
 {
-    int i, c, m, l = 0, r = (int)recdef->n_fields - 1;
+    int i, c, m, l = 0, r = recdef->n_fields - 1;
     while (l <= r) {
         m = (l + r) / 2;
         i = recdef->sorted_fields[m];
@@ -650,7 +650,7 @@ int lookup_record_field(struct b_constructor *recdef, dptr query, struct inline_
          * Check if we have a inline cache match.
          */
         if (ic->class == (union block *)recdef)
-            return (int)ic->index;
+            return ic->index;
 
         fnum = IntVal(*query);
 
@@ -678,7 +678,7 @@ int lookup_record_field(struct b_constructor *recdef, dptr query, struct inline_
              */
             word i = cvpos(IntVal(*query), nf);
             if (i != CvtFail && i <= nf)
-                return (int)i - 1;
+                return i - 1;
             else
                 return -1;
         }
